Create "Factorize"
This program is all about the factors of numbers.  The main method returns an array of the number that is passed in.  The other methods use this one to figure out if a number is perfect, find the greatest common factor between two numbers, and find the least common multiple between two factors.